Make chromium style checker happy about PlusProfile
Due to its three std::string members, PlusProfile is considered a
complex struct by the chromium style checker. This forces us to define
an explicit out-of-line constructor.

Currently, a hacky template is used to circumvent this check, since a
lot of the code needs to be migrated away from aggregate initialisation
to make a constructor work. This CL does so, by defining a constructor
for PlusProfile which accepts all for members of the struct.

Outside of tests, rewriting the initialisation is straightforward, since
most places set all four members anyways.

Inside tests, the PlusProfiles are often incomplete. In these cases, the
CL adapts them to use test::CreatePlusProfile[2]() where possible.
In a few tests, this creates some awkwardness, because
PlusAddressService::SavePlusAddress(url::Origin, PlusProfile) takes an
origin and a profile, even though PlusProfile::facet should already be
the origin. The function's implementation simply ignores
PlusProfile::facet and saves a PlusProfile for the passed in origin
instead.
Since the function ignores the facet, it is not necessary to enforce
that the profile's facet matches the origin in tests.
Longer term, we should just get rid of the origin parameter, though.

Bug: b/322279583, b/322147254
Change-Id: I7db29d129f1cb564bbbbb30681744d3fb6dfaa88
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5453442
Reviewed-by: Jan Keitel <jkeitel@google.com>
Commit-Queue: Florian Leimgruber <fleimgruber@google.com>
Cr-Commit-Position: refs/heads/main@{#1288582}