[Uno-d] Profiles created through signed in intercept remains signed in
The sign in intercept will make it clear that the user will be signed in
to chrome in the new profile. Therefore, the user is considered as
explicitly signed in. Declining sync will no longer clear the primary
account.

Note: UI strings will be updated as soon as they are finalized.

Bug: b/319409641
Change-Id: I872d8bc51a7686c104575b4cd9784a1da803ac2b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5458592
Reviewed-by: Ryan Sultanem <rsult@google.com>
Commit-Queue: Monica Basta <msalama@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1288595}