Load native library through an instrumentation key for Cronet in AOSP only
Currently, Cronet runs native tests in AOSP, in order to do so, there is
a translation layer which translates GN build rules to Soong modules. However,
this translation layer is currently incapable of removing the NativeLibraries
that is generated from the intermediate targets and replacing it with the real
NativeLibraries that is generated at the root of the build tree. In order to
go around that, we decided to use an instrumentation key which seems to be working
well for us. This CL basically downstreams the changes that we have in AOSP to
Chromium to reduce divergence.

I believe that the loading of the native libraries should be done through the
instrumentation key for the entire of Chromium as that will speed-up the compilation
process and simplify it. However, I have not investigated the feasibility of that.

From Chromium perspective, this change should be a no-op as that flag is always
null. However, from AOSP, we will be benefitting from that a lot :D

Bug: 334809455
Change-Id: Ibd3e111d271c846ba7c95f95ef1e7f56d514a714
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5458753
Reviewed-by: Andrew Grieve <agrieve@chromium.org>
Auto-Submit: Mohannad Farrag <aymanm@google.com>
Commit-Queue: Mohannad Farrag <aymanm@google.com>
Cr-Commit-Position: refs/heads/main@{#1288614}