Remove PDM friend declarations from ADM and PayDM
The ADM and PayDM currently befriend the PDM. This was a temporary hack
that made moving code easier, so not all dependencies needed to move at
once. Now, pretty much all the code has moved and the friend
declarations are removed by this CL.

This CL:
- Adds a getter for ADM::has_initial_load_finished_ and
  PayDM::is_payments_data_loaded_, since PDM::IsDataLoaded() accesses
  them.
- Makes PayDM::HasPendingPaymentQueries() public, since
  PDM::NotifyPersonalDataObserver() calls it.
- Makes PayDM::pending_(creditcards|ibans|...)_query_ protected instead
  of private, so the TestPayDM friend declaration can be removed from
  PayDM too.

A TestPDM friend declaration remains in the PayDM. Removing it will
require moving some TestPDM code to TestPayDM and is deferred to a
follow-up.

Bug: b/322170538
Change-Id: I6adb0e9f9b7d950850e920cf2ca7caab9cb3146d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5458432
Reviewed-by: Jan Keitel <jkeitel@google.com>
Commit-Queue: Florian Leimgruber <fleimgruber@google.com>
Cr-Commit-Position: refs/heads/main@{#1288634}