FLEDGE: Move most header checks in AuctionDownloader to OnResponseStarted.
We want to check Ad-Auction-Allow-Trusted-Scoring-Signals-From at that
time as it needs to gate cross-origin trusted signals fetch, and it
would be weird to have it permit the fetch even when Ad-Auction-
Allowed was not set, or if it was set on a 404, etc.

Sadly the charset check can't actually happen this early (since it
looks at the body); and so the mimetype check is kept with it since it
would be strange to spec parsing out content-type twice.

Bug: 332913415
Change-Id: I0ed69f54fbf5490a4557b5dd55e579410cff3913
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5454594
Reviewed-by: mmenke <mmenke@chromium.org>
Commit-Queue: Maks Orlovich <morlovich@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1288677}