[bookmarks] Broadcast base::Location triggering deletions
The general ideal is that BookmarkModelObserver should be able to
provide details about which piece of code (base::Location) was
responsible for issuing a deletion.

The only planned consumer for this location is
sync_bookmarks::BookmarkModelObserverImpl. The logic there isn't
implemented yet but TODOs are added to convert this location to
something that can be exposed to Sync code and consequently to the
Sync server.

Change-Id: If679ec259c879da2318e3297ab9d4d1301e30c11
Bug: 334001702
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5445089
Reviewed-by: Scott Violet <sky@chromium.org>
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Reviewed-by: Matthew Jones <mdjones@chromium.org>
Commit-Queue: Mikel Astiz <mastiz@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1288683}