Backlight A11y: AX trees from the OCR Service come with no tree ID and tree data
When trying to test Backlight accessibility end-to-end, I ran across
the issue that the tree update received from the
Screen AI Service has no tree data.
This should not happen, but since would require changes to other
parts of the code, and since we already set our own
tree ID, I simply removed the check and moved it to a TODO.

R=kyungjunlee@google.com

AX-Relnotes: n/a.
Bug: b/303133098
Change-Id: I3c75e3c7245357c53a9183742504d8f2b26c0148
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5457780
Reviewed-by: Kyungjun Lee <kyungjunlee@google.com>
Commit-Queue: Nektarios Paisios <nektar@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1288701}