[waffle]Change button state when needed on resize
The "More" button is displayed when the list of search engines doesn't
fit in the screen. The button should change to "Set as default" when
the screen is resized enough to fit the whole content.

We also re-enable the SearchEngineChoiceUIPixelTest pixel tests in this
change and fix a flakiness that was happening due to a hover color.

Recording: https://crbug.com/333787236#comment5
Bug: b:333787236
Change-Id: Ief1fd62f43a88a2b391c0d3e5f5d783d860b0364
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5453877
Commit-Queue: Jack Yammine <jyammine@google.com>
Reviewed-by: Nicolas Dossou-Gbété <dgn@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1288703}