LayoutObject: might_traverse_physical_fragments_=true initially.
Instead of false.

This flag is updated whenever the object is inserted into the layout
tree, so, as long as an object actually IS inserted into the tree, this
change doesn't make any difference.

LayoutView used to manually set this flag (since it is the root and is
therefore not inserted into the tree). No longer needed.

As for other objects not inserted into the tree... the only current case
for now should be the LayoutBlockFlow created by FormattedText. There
are no reasons why this one shouldn't be fragment-traversed, though.

There are some upcoming changes on the pagination side, with more
non-inserted layout objects. They should also be fragment-traversed. See
crrev.com/c/5453623 , and this is the motivation behind this change.

Change-Id: Icf00a4c7cf896771b8a6787a84409ed84d194183
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5458295
Commit-Queue: Morten Stenshorne <mstensho@chromium.org>
Reviewed-by: Ian Kilpatrick <ikilpatrick@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1288744}