[Product specifications] Prevent internals crash when sync flag off
It's possible for the product specifications service to be null due to
the added flag specifically for the sync part of the feature. This
patch adds an early exit for that case.

Bug: b:335415856
Change-Id: I90e7f03ab4a036d2b5912e867bf5b440ed1a0782
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5461378
Reviewed-by: Ayman Almadhoun <ayman@chromium.org>
Commit-Queue: Ayman Almadhoun <ayman@chromium.org>
Auto-Submit: Matthew Jones <mdjones@chromium.org>
Commit-Queue: Matthew Jones <mdjones@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1288749}