iOS: Offer account switch when tapping on the NTP avatar
This CL adds the option, behind a feature flag, to show an
account-switching UI when tapping the NTP identity disc while signed
in (previously/with the feature disabled, this opens Chrome settings).
For this initial version, the "account switching UI" is in fact just
the existing "InstantSignin" UI, which happens to support
account-switching just fine.

Bug: 334867766
Change-Id: Iaf0c94ec109e75648f557924fbddc45d286d73a4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5444621
Reviewed-by: Sylvain Defresne <sdefresne@chromium.org>
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Commit-Queue: Marc Treib <treib@chromium.org>
Reviewed-by: Arthur Milchior <arthurmilchior@chromium.org>
Reviewed-by: Robert Kaplow <rkaplow@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1288793}