Permit individual success/failure per OS registration item
While 3 of the 4 Android APIs for registration are atomic (all items
succeed or all fail), 1 of the 4 is not, and therefore we need to return
a list of results rather than just 1. The outer callback is now
(correctly) invoked upon the last JNI call returning, whereas before
it was the first.

Bug: 334755569
Change-Id: I79a785a30c29d900fdec03c4664e25ed67bfe39d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5457979
Commit-Queue: Andrew Paseltiner <apaseltiner@chromium.org>
Reviewed-by: John Delaney <johnidel@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1288796}