[STGV2] Cleanup CreateHistoricalGroup and group closing.
TabRestore and SavedTabGroups both need to perform actions when the
group is closing in the tabstrip. These operations need to be performed
in a specific order. CreateHistoricalGroup was not the right place to
perform these operations since it can be called in more places than
when the group is closing. Because of this a new function is added...
WillCloseGroup that performs all of the operations that should be
applied when a tab group is closing.



Change-Id: If8efe8ee6b96696f7f96d15fe321488ce1c330b6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5459934
Reviewed-by: Darryl James <dljames@chromium.org>
Commit-Queue: David Pennington <dpenning@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1288830}