Separate USD from FRE
Unified State Determination (USD) is really about determining state.
When USD is on, a device will first check if it is of interest to Google
(meaning it's either managed or destined to be managed) using a private
protocol (PSM), then retrieve a state indicating how to proceed during
OOBE from a management perspective, if it is of interest.

Forced Re-Enrollment (FRE) also does a state determination, but using a
separate mechanism (the hash dance). That mechanism relies on state
keys, which are not supported on all platforms (notably, Flex right
now, while FRE on Flex is under development).

Today's code conflate "USD is enabled," which is now true on Flex, and
"FRE is enabled," which actually means FRE is supported, and triggers
the use of state keys. This creates an obvious conflict and breakage on
Flex.

This CL attempts to untangle that mess, so that one can have USD
separate from state keys, which truly is one of the main goals of USD
in the first place!

We also disable legacy FRE on Flex, because we don't have a mechanism to
record whether a device is managed across power wash and don't want to
use a non private protocol blindly on consumer devices.

Bug: b:314539401
Test: unit_tests --gtest_filter=*AutoEnrollmentTypeChecker*Test*
Change-Id: Ib37845bb7b3c6d39e3b890898ddea7be978144f0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5460586
Commit-Queue: Yves Arrouye <drcrash@chromium.org>
Auto-Submit: Yves Arrouye <drcrash@chromium.org>
Reviewed-by: Sergiy Belozorov <sergiyb@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1288833}