[PWA/Universal Install] Change webcontents observers for dialogs
This CL does the following:
1. Following feedback in crrev.com/c/5444573, renames some of the
classes for the create shortcut flow to be called create desktop
shortcut to be more in alignment with what it is supposed to do. This
also helps differentiate between the current create shortcut dialog
code and the new one a bit more.

2. Updates the WebContents observers to listen to PrimaryPageChanged()
instead of DidFinishNavigation(), which is the preferred way as per
the comments in web_contents_observer.h.

Essentially, this CL is a no-op, and has been verified by both
manually testing the 4 dialogs (simple, detailed, DIY and new create
shortcut) as well as running the existing automated tests to verify
functionality changes has not been introduced by mistake.

Bug: b/335202700, crbug/333024272
Change-Id: Iff18a50a58984ec6e5424bd1f5c3fd209aac3d20
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5460401
Reviewed-by: Mike Wasserman <msw@chromium.org>
Commit-Queue: Dibyajyoti Pal <dibyapal@chromium.org>
Reviewed-by: Daniel Murphy <dmurph@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1288861}