[Autofill] Disable TestCardUnmaskPromptWaiter in captured-site tests
The captured-site-test bot is currently broken because of a nullptr
dereference in TestCardUnmaskPromptWaiter.

The waiter is currently not needed because there are no actual tests
that bring up a card unmask prompt.

Therefore, this CL removes the TestCardUnmaskPromptWaiter for the
time being to make the bot green again.

Bug: 333815150
Change-Id: Ia6e0fad688bca2aa90a24fc8d9779a27bd1185a8
Cq-Do-Not-Cancel-Tryjobs: true
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5458703
Reviewed-by: Ian Struiksma <ianstruiksma@google.com>
Reviewed-by: Siyu An <siyua@chromium.org>
Commit-Queue: Christoph Schwering <schwering@google.com>
Cr-Commit-Position: refs/heads/main@{#1288877}