[PWA] Relax check in WebAppUninstallAndReplaceJob for existing app
Previously, the WebAppUninstallAndReplaceJob would DCHECK and fail if
WebAppRegistrar::IsInstalled() is false. However, as noticed in the
linked bug, this caused SWAs to run into a crash loop, since an app
marked for uninstallation fails the previous CHECK.

This CL loosens that DCHECK to only verify if a web app exists in the
web app registry. The DCHECK is also converted to a CHECK to keep
up with current chromium dev guidelines

Bug: 335253048
Change-Id: Ifee261eefcbba848dd133dd7dc26a4502f5608a0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5462477
Reviewed-by: Daniel Murphy <dmurph@chromium.org>
Commit-Queue: Dibyajyoti Pal <dibyapal@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1288942}