Enable Safety Hub by default
To be able to enable by default, this CL also makes some small changes.
Such as:
- PasswordStatusCheckService will be created by factory only if there is
a profile store available.The profile store may not be available for OTR
profile and some tests.
- SafetyHubMenuNotificationService will only add password notification
if PasswordStatusCheckService is not null.
- PasswordStatusCheckService and NotificationPermissionReviewService
will not be create with browser context anymore. This was not necessary
and was causing some test to be failed whenever profile store is not
available.

Bug: b/40267370
Change-Id: I08fc740de2765b4e82d73858d055f7800b0fe396
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5454291
Commit-Queue: Side YILMAZ <sideyilmaz@chromium.org>
Reviewed-by: Martin Šrámek <msramek@chromium.org>
Reviewed-by: Devlin Cronin <rdevlin.cronin@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1288949}