CrOS Settings: Fix crash due to missing multitasking string key
Opening Settings immediately crashes due to a failed CHECK where a
key in $i18n{} is not properly registered. The missing key is
"snapWindowLabel" which is not registered when the
OsSettingsRevampWayfinding and FasterSplitScreenSetup features are
both disabled.

This CL ensures the strings are always registered in load time data
regardless of the feature states. The feature feature visibility is
then controlled on the frontend.

Bug: b:334223004
Test: browser_tests --gtest_filter="*OSSettings*Personalization*"
Change-Id: I1a17c7dcd2315b6cdea39854f973c21a3f5e48fc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5463266
Commit-Queue: Wes Okuhara <wesokuhara@google.com>
Reviewed-by: Michele Fan <michelefan@chromium.org>
Reviewed-by: Sophie Wen <sophiewen@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1288957}