Disable child accounts from toggling consent in privacy setting page.
The child account consent is now independent of the device owners
consent. Parents are given the ability to control the UMA consent
in the OOBE flow (parent handoff).

As children are not allowed to consent meaningfully, as an interim
solution, disable the option for child accounts to toggle UMA consent
in the privacy settings page. This means that once a parent sets the
UMA consent in the OOBE flow for the child account, it cannot be
updated.

Consent Flow Redesign: go/cros-per-user-consent-flow-changes

BUG=283174830

Change-Id: I8dc28cceb6f47b0b743fe1352bff51c82386ec8d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5459562
Commit-Queue: Hirthanan Subenderan <hirthanan@google.com>
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1288960}