[views_ax] Introduce ComputeIsIgnored to ViewAccessibility
This CL introduces some changes to how we add the `kIgnored` state
for a View.

Previously, this state would be added either by calling `SetIsIgnored`
or added to the out `data` in `GetAccessibleNodeData` in one of the
last steps of the function.

Now, `SetIsIgnored` sets a private member in ViewAccessibility, and
then it calls a new method introduced by this CL,
`ComputeIgnoredState`. This function will either add or remove the
`kIgnored` state according to three different conditions, whether
the view is set to be ignored by the new member, or if it has been
marked as `pruned_` or if the role is `kNone`
(this is AXNodeData behavior). `ComputeIgnoredState` is now called
whenever either of these conditions change, in order to keep the
state up to date.

This is so that we clean up GetAccessibleNodeData and move in the
direction of having the AXNodeData state be updated via "push"
rather than "pull".


Bug: 325137417
Change-Id: I39d11c7bf3aa997b7abee5ebf12e6cce7649b54a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5463260
Commit-Queue: Javier Contreras <javiercon@microsoft.com>
Reviewed-by: Benjamin Beaudry <benjamin.beaudry@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1289024}