[Default Nav Transition] Screenshot same-doc navs: viz changes
Implements the changes in viz: issue the CopyOutputRequest when receive
the CompositorFrame with a destination token, and sends the
CopyOutputResult back to the browser when it's ready.

When the browser receives the DidCommit message, a callback is set to
the HostFrameSinkManager to instruct where to stash the screenshot when
Viz sends back the screenshot.

Low-Coverage-Reason: TESTS_IN_SEPARATE_CL browser tests in the last CL
Bug: 40259037
Change-Id: I7d9d8fff36405312914f24c45670b6301450f0d1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5445171
Commit-Queue: William Liu <liuwilliam@chromium.org>
Reviewed-by: Khushal Sagar <khushalsagar@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1289028}