[PWA] Allow WebAppSyncBridge to uninstall all types of web apps
Previously, the WebAppSyncBridge would only uninstall web apps that
were marked as |is_uninstalling|=true if the web app was user
uninstallable. However, this could cause a crash loop on ChromeOS
for management types like kSystem where a faulty SWA installation
would fail on startup if it was previously not uninstalled
properly.

This CL fixes that, by introducing a destructive version of the uninstall flow that is called RemoveAllManagementTypesAndUninstall(),
that is passkeyed to the WebAppSyncBridge to prevent other public interfaces from directly calling into it.

Also adds a unit test verifying that system web apps can really be uninstalled on startup if is_uninstalling is set.

Bug: 335253048
Change-Id: If7a0a0bc0610750b863e7f1649bc1b7e36ab9d8c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5462473
Commit-Queue: Dibyajyoti Pal <dibyapal@chromium.org>
Reviewed-by: Daniel Murphy <dmurph@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1289053}