Delete AppShimMenuController.
This class was used to change out Chrome's menu bar when a legacy
packaged app was activated in Chrome. Legacy packaged apps are no
longer supported on macOS except with some exceptions for
development purposes, and for those cases it shouldn't matter what
the menu bar is showing.

This avoids the need to fix flaky tests for this class.

Fixed: 41486359, 40936646
Change-Id: Iea798e87008bcaf5c4c6a0e025df7cac4bccd75a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5464166
Reviewed-by: Avi Drissman <avi@chromium.org>
Commit-Queue: Marijn Kruisselbrink <mek@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1289059}